reorganizes file structure to better seperate front and back ends
