+++ src/client/js/toggleManyOptions.js
