+++ src/client/js/toggleOptionViz.js
