+++ src/data/equipment.js
