+++ src/data/equipment.json
