+++ src/data/spells.js
