+++ src/data/spells.json
