+++ src/data/weapons.csv
