+++ src/server/index.js
@@ -2,11 +2,11 @@ const path = require('path');
 const express = require('express');
 const app = express();
 const { body, validationResult } = require('express-validator');
-const generator = require('./js/generators.js');
+const generator = require('./generators.js');
 
 //dev requires
-const interpreter = require('./js/interpreter.js');
-const prettify = require('./js/characterpage.js');
+const interpreter = require('../client/js/interpreter.js');
+const prettify = require('../client/js/characterpage.js');
 
 const port = process.env.PORT || 3000;
 
@@ -14,11 +14,11 @@ app.use(express.static(__dirname));
 app.use(require('body-parser').urlencoded({ extended: false }));
 
 app.get('/', (req, res) => {
-    res.status(200).sendFile(path.join(__dirname, '/index.html'));
+    res.status(200).sendFile(path.join(__dirname, '../client', '/index.html'));
 })
 
 app.get('/custom', (req, res) => {
-    res.status(200).sendFile(path.join(__dirname, '/custom.html'));
+    res.status(200).sendFile(path.join(__dirname, '../client', '/custom.html'));
 });
 
 app.post('/', (req, res) => {
